+++ R/gather_files_names.R
@@ -4,17 +4,20 @@
 #'
 #' @keywords internal
 #' @param paths Character vector of file paths.
-#' @returns Character vector of file names.
+#' @param verbose Print messages.
+#' @returns Named character vector. 
 #' @inheritParams gather_files
 #' @importFrom stringr str_split
+#' @importFrom stats setNames
 gather_files_names <- function(paths,
                                type,
-                               nfcore_cutandrun){
-  message("Constructing file names.")
+                               nfcore_cutandrun,
+                               verbose=TRUE){
+  messager("Constructing file names.",v=verbose)
   if(isTRUE(nfcore_cutandrun)){
     if(startsWith(type,"peaks")){
       # paths <- grep("03_peak_calling", paths, value = TRUE)
-      names <- paste(
+      list_names <- paste(
         basename(
           dirname(
             dirname(
@@ -28,7 +31,7 @@ gather_files_names <- function(paths,
                            simplify = TRUE)[,1],
         sep=".")
     } else if(type=="picard"){
-      names <- paste(
+      list_names <- paste(
         basename(
           dirname(
             dirname(
@@ -47,9 +50,39 @@ gather_files_names <- function(paths,
         stringr::str_split(basename(paths),"[.]",
                            simplify = TRUE)[,1],
         sep=".")
+    } else if(type=="multiqc"){
+        list_names <- paste(
+            basename(dirname(dirname(paths))),
+            stringr::str_split(basename(paths),"[.]",
+                               simplify = TRUE)[,1],
+            sep="."
+        ) 
+    }else if(startsWith(type,"bowtie")){
+        list_names <- paste(
+            basename(dirname(dirname(dirname(dirname(paths))))),
+            stringr::str_split(basename(paths),"[.]",
+                               simplify = TRUE)[,1],
+            sep="."
+        ) 
+    }else if(type=="trimgalore"){
+        list_names <- paste(
+            basename(dirname(dirname(dirname(paths)))),
+            stringr::str_split(basename(paths),"[.]",
+                               simplify = TRUE)[,1],
+            sep="."
+        )
+    }else if(startsWith(type,"bam")){
+        list_names <- paste(
+            basename(dirname(dirname(dirname(dirname(paths))))),
+            stringr::str_split(basename(paths),"[.]",
+                               simplify = TRUE)[,1],
+            sep="."
+        )
+    }else{
+        list_names <- make.unique(basename(paths))
     }
   } else {
-    names <- make.unique(basename(paths))
+    list_names <- make.unique(basename(paths))
   }
-  return(names)
+  return(stats::setNames(paths,list_names))
 }