+++ R/read_multiqc.R
@@ -1,0 +1,6 @@
+read_multiqc <- function(path,
+                         verbose=TRUE){
+    
+    messager("Reading multiQC report.",v=verbose)
+    data.table::fread(path)
+}